rename(view): views->view 폴더명 변경
